+++ demo/admin/src/products/future/CreateCapProductPage.tsx
@@ -8,6 +8,7 @@ import {
     ToolbarBackButton,
     ToolbarFillSpace,
 } from "@comet/admin";
+import { ContentScopeIndicator } from "@comet/cms-admin";
 import { CreateCapProductForm } from "@src/products/future/generated/CreateCapProductForm";
 import * as React from "react";
 import { useIntl } from "react-intl";
@@ -17,7 +18,7 @@ export function CreateCapProductPage(): React.ReactElement {
     return (
         <Stack topLevelTitle={intl.formatMessage({ id: "products.createCapProduct", defaultMessage: "Create Cap Product" })}>
             <SaveBoundary>
-                <StackToolbar>
+                <StackToolbar scopeIndicator={<ContentScopeIndicator global />}>
                     <ToolbarBackButton />
                     <ToolbarAutomaticTitleItem />
                     <ToolbarFillSpace />