+++ packages/admin/cms-admin/src/generator/future/generateGrid.ts
@@ -299,15 +299,14 @@ export function generateGrid(
     const code = `import { gql, useApolloClient, useQuery } from "@apollo/client";
     import {
         CrudContextMenu,
+        DataGridToolbar,
         filterByFragment,
         GridFilterButton,
         GridColDef,
         muiGridFilterToGql,
         muiGridSortToGql,
         StackLink,
-        Toolbar,
         ToolbarActions,
-        ToolbarAutomaticTitleItem,
         ToolbarFillSpace,
         ToolbarItem,
         useBufferedRowCount,
@@ -408,8 +407,7 @@ export function generateGrid(
         toolbar
             ? `function ${gqlTypePlural}GridToolbar() {
         return (
-            <Toolbar>
-                <ToolbarAutomaticTitleItem />
+            <DataGridToolbar>
                 ${
                     hasSearch
                         ? `<ToolbarItem>
@@ -434,7 +432,7 @@ export function generateGrid(
                 </ToolbarActions>`
                         : ""
                 }
-            </Toolbar>
+            </DataGridToolbar>
         );
     }`
             : ""
@@ -609,8 +607,8 @@ export function generateGrid(
                 ${
                     toolbar
                         ? `components={{
-Toolbar: ${gqlTypePlural}GridToolbar,
-}}`
+                            Toolbar: ${gqlTypePlural}GridToolbar,
+                        }}`
                         : ""
                 }
             />