+++ packages/admin/admin-react-select/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @comet/admin-react-select
 
+## 7.0.0-beta.2
+
+### Patch Changes
+
+-   Updated dependencies [2fc764e29]
+    -   @comet/admin@7.0.0-beta.2
+    -   @comet/admin-icons@7.0.0-beta.2
+
 ## 7.0.0-beta.1
 
 ### Patch Changes
@@ -156,6 +164,15 @@
 -   Updated dependencies [b0249e3bc]
 -   Updated dependencies [92eae2ba9]
     -   @comet/admin@7.0.0-beta.0
+
+## 6.13.0
+
+### Patch Changes
+
+-   Updated dependencies [5e25348bb]
+-   Updated dependencies [796e83206]
+    -   @comet/admin@6.13.0
+    -   @comet/admin-icons@6.13.0
 
 ## 6.12.0